Adicionando alterações ao repositório

+++ PlanilhasJava/src/main/java/br/fag/estimativa/Estimativa.java
@@ -11,5 +11,5 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Estimativa extends Concurso {
-    private Float estimativa;
+    private Integer estimativa;
 }
